move validDomains tests into existing manifest-builder-icons test file
